finally implemented the tri-modal serialization, and potentially also factories. Now everything need testing.
