--- UnityPlus/Assets/_UnityPlus.Serialization/PersistentModeInterface.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections;
-using UnityEngine;
-
-namespace UnityPlus.Serialization
-{
-    public class PersistentModeInterface
-    {
-
-    }
-}
